requirement.txt
streamlit
numpy
pillow
rembg
scikit-learn
tensorflow
onnxruntime
protobuf
